Final release!
Unfortunately we will stop developing the game because no one in the team is really useful except for a person so it would be a finger in the ass to finsh it with all the features needed. However the game is still AMAZING as fuck and it will be surely funny to play between friends to try a different version of black jack. You will compete against each others without fish with everything based on how many times you win. Probably tomorrow there will be another update with some bug fix and maybe I will start to translate the game from italian to english. Stay tuned folks!